# munayfund-api2/docs/swagger.yaml
basePath: /api/v1
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  httputil.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
...
  munayfund-api2_internal_core_domain.Multimedia:
    properties:
      type:
        type: string
      url:
        type: string
    type: object
  munayfund-api2_internal_core_domain.Project:
    properties:
      advances:
        items:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.ProjectAdvance'
        type: array
      description:
        type: string
      id:
        type: string
      multimedia:
        items:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.Multimedia'
        type: array
      title:
        type: string
      userId:
        type: string
    type: object
  munayfund-api2_internal_core_domain.ProjectAdvance:
    properties:
      date:
        type: string
      details:
        type: string
      media:
        items:
          $ref: '#/definitions/munayfund-api2_internal_core_domain.Multimedia'
        type: array
    type: object
...
host: localhost:8085
info:
  contact:
    email: todefine@github.com
    name: KaniKrafters
    url: https://todefine.github.com
  description: This API is in charge of leveraging some functionalities of munayFund
    a next gen crowdfunding solution.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Swagger MunayFund API WEB 2
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de proyectos.
      parameters:
      - default: 10
        description: Número máximo de proyectos por página
        in: query
        name: limit
        type: integer
      - default: 1
        description: Número de página
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de proyectos
          schema:
            items:
              $ref: '#/definitions/munayfund-api2_internal_core_domain.Project'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Obtener lista de proyectos
      tags:
      - projects
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina un proyecto por su ID.
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del proyecto a eliminar
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Proyecto eliminado exitosamente
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Eliminar proyecto
      tags:
      - projects
    get:
      consumes:
      - application/json
      description: Obtiene los detalles de un proyecto específico.
      parameters:
      - description: ID del proyecto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detalles del proyecto
          schema:
            $ref: '#/definitions/munayfund-api2_internal_core_domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Obtener detalles del proyecto
      tags:
      - projects
  /projects/{id}/media:
    put:
      consumes:
      - multipart/form-data
      description: Este endpoint permite actualizar la multimedia de un proyecto existente.
      operationId: updateProjectMultimedia
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del proyecto
        in: path
        name: id
        required: true
        type: string
      - description: Archivos multimedia a agregar
        in: formData
        name: files
        required: true
        type: file
      - description: Lista de objetos multimedia actualizados
        in: body
        name: updatedMultimedia
        required: true
        schema:
          items:
            $ref: '#/definitions/munayfund-api2_internal_core_domain.Multimedia'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: Proyecto actualizado exitosamente
          schema:
            $ref: '#/definitions/munayfund-api2_internal_core_domain.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Actualiza la multimedia de un proyecto.
      tags:
      - projects
  /projects/create:
    post:
      consumes:
      - multipart/form-data
      description: Crea un nuevo proyecto con información multimedia.
      operationId: create-project
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID del usuario creador
        in: formData
        name: userId
        required: true
        type: string
      - description: Título del proyecto
        in: formData
        name: title
        required: true
        type: string
      - description: Descripción del proyecto
        in: formData
        name: description
        required: true
        type: string
      - description: Archivos multimedia (imágenes y videos)
        in: formData
        name: media
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Operación exitosa
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - ApiKeyAuth: []
      summary: Crear un nuevo proyecto
      tags:
      - projects
...
  /search:
    get:
      consumes:
      - application/json
      description: Obtiene la lista de proyectos cuyos nombres son similares al proporcionado.
      parameters:
      - description: Nombre parcial del proyecto a buscar
        in: query
        name: partialName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de proyectos
          schema:
            items:
              $ref: '#/definitions/munayfund-api2_internal_core_domain.Project'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Obtener proyectos por nombre similar
      tags:
      - projects
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"